chore: rename optional dependency helper
