+++ pandasai/helpers/optional.py
